refactor(tui): extract keymap picker integration
Move the ChatWidget-specific /keymap picker entry points into a dedicated
chatwidget module while keeping picker construction and edit semantics in
keymap_setup.

Document the synchronization contract between persisted keymap config,
app-level shortcut caches, and bottom-pane runtime bindings.
