+++ codex-rs/tui/src/chatwidget.rs
@@ -114,7 +114,6 @@ use codex_chatgpt::connectors;
 use codex_config::ConfigLayerStackOrdering;
 use codex_config::types::ApprovalsReviewer;
 use codex_config::types::Notifications;
-use codex_config::types::TuiKeymap;
 use codex_config::types::WindowsSandboxModeToml;
 use codex_core_skills::model::SkillMetadata;
 use codex_features::FEATURES;
@@ -376,7 +375,6 @@ use crate::key_hint::KeyBinding;
 use crate::key_hint::KeyBindingListExt;
 use crate::keymap::ChatKeymap;
 use crate::keymap::RuntimeKeymap;
-use crate::keymap_setup;
 use crate::render::Insets;
 use crate::render::renderable::ColumnRenderable;
 use crate::render::renderable::FlexRenderable;
@@ -396,6 +394,7 @@ use self::goal_status::goal_status_indicator_from_app_goal;
 mod goal_menu;
 mod interrupts;
 use self::interrupts::InterruptManager;
+mod keymap_picker;
 mod session_header;
 use self::session_header::SessionHeader;
 mod skills;
@@ -8300,97 +8299,6 @@ impl ChatWidget {
         self.bottom_pane.show_selection_view(params);
     }
 
-    fn open_keymap_picker(&mut self) {
-        match RuntimeKeymap::from_config(&self.config.tui_keymap) {
-            Ok(runtime_keymap) => {
-                let params = keymap_setup::build_keymap_picker_params(
-                    &runtime_keymap,
-                    &self.config.tui_keymap,
-                );
-                self.bottom_pane.show_selection_view(params);
-            }
-            Err(err) => {
-                self.add_error_message(format!("Invalid `tui.keymap` configuration: {err}"));
-            }
-        }
-    }
-
-    pub(crate) fn open_keymap_action_menu(
-        &mut self,
-        context: String,
-        action: String,
-        runtime_keymap: &RuntimeKeymap,
-    ) {
-        let params = keymap_setup::build_keymap_action_menu_params(
-            context,
-            action,
-            runtime_keymap,
-            &self.config.tui_keymap,
-        );
-        self.bottom_pane.show_selection_view(params);
-    }
-
-    pub(crate) fn open_keymap_capture(
-        &mut self,
-        context: String,
-        action: String,
-        intent: crate::app_event::KeymapEditIntent,
-        runtime_keymap: &RuntimeKeymap,
-    ) {
-        let view = keymap_setup::build_keymap_capture_view(
-            context,
-            action,
-            intent,
-            runtime_keymap,
-            self.app_event_tx.clone(),
-        );
-        self.bottom_pane.show_view(Box::new(view));
-        self.request_redraw();
-    }
-
-    pub(crate) fn open_keymap_replace_binding_menu(
-        &mut self,
-        context: String,
-        action: String,
-        runtime_keymap: &RuntimeKeymap,
-    ) {
-        let params =
-            keymap_setup::build_keymap_replace_binding_menu_params(context, action, runtime_keymap);
-        self.bottom_pane.show_selection_view(params);
-    }
-
-    pub(crate) fn return_to_keymap_picker(
-        &mut self,
-        context: &str,
-        action: &str,
-        runtime_keymap: &RuntimeKeymap,
-    ) {
-        let params = keymap_setup::build_keymap_picker_params_for_selected_action(
-            runtime_keymap,
-            &self.config.tui_keymap,
-            context,
-            action,
-        );
-        let replaced = self.bottom_pane.replace_active_views_with_selection_view(
-            &[
-                keymap_setup::KEYMAP_PICKER_VIEW_ID,
-                keymap_setup::KEYMAP_ACTION_MENU_VIEW_ID,
-                keymap_setup::KEYMAP_REPLACE_BINDING_MENU_VIEW_ID,
-            ],
-            params,
-        );
-        if !replaced {
-            let params = keymap_setup::build_keymap_picker_params_for_selected_action(
-                runtime_keymap,
-                &self.config.tui_keymap,
-                context,
-                action,
-            );
-            self.bottom_pane.show_selection_view(params);
-        }
-        self.request_redraw();
-    }
-
     fn status_line_context_window_size(&self) -> Option<i64> {
         self.token_info
             .as_ref()
@@ -12158,24 +12066,6 @@ impl ChatWidget {
     /// runtime overrides applied via TUI, e.g., model or approval policy).
     pub(crate) fn config_ref(&self) -> &Config {
         &self.config
-    }
-
-    pub(crate) fn apply_keymap_update(
-        &mut self,
-        keymap_config: TuiKeymap,
-        runtime_keymap: &RuntimeKeymap,
-    ) {
-        self.config.tui_keymap = keymap_config;
-        self.copy_last_response_binding = runtime_keymap.app.copy.clone();
-        self.chat_keymap = runtime_keymap.chat.clone();
-        self.queued_message_edit_hint_binding = queued_message_edit_hint_binding(
-            &self.chat_keymap.edit_queued_message,
-            terminal_info(),
-        );
-        self.bottom_pane
-            .set_queued_message_edit_binding(self.queued_message_edit_hint_binding);
-        self.bottom_pane.set_keymap_bindings(runtime_keymap);
-        self.request_redraw();
     }
 
     #[cfg(test)]
